refactor: refactoring to use new eslint file format
Does not yet work.
