+++ eslint.config.js
@@ -1,0 +1,5 @@
+const index = require('./dist/');
+
+console.log(index);
+
+module.exports = index.default;
