+++ package.json
@@ -3,10 +3,11 @@
   "version": "0.0.0-development",
   "description": "The ultimate oppressive style guide.",
   "scripts": {
-    "lint": "eslint -c index.js test/example.ts",
-    "lint:mild": "eslint -c mild.js test/example.ts",
-    "test": "eslint -c index.js test/good-file.ts",
-    "test:mild": "eslint -c mild.js test/good-file.ts",
+    "build": "rm -rf dist && tsc -p ./tsconfig.build.json",
+    "lint": "eslint test/example.ts",
+    "lint:mild": "eslint test/example.ts",
+    "test": "eslint test/good-file.ts",
+    "test:mild": "eslint test/good-file.ts",
     "prepare": "husky install",
     "semantic-release": "semantic-release"
   },
@@ -20,19 +21,10 @@
   },
   "main": "./index.js",
   "license": "ISC",
-  "files": [
-    "index.js",
-    "jest.js",
-    "react.js",
-    "mild.js",
-    "prettier.json",
-    "*.md",
-    "rules/**",
-    "helpers/**"
-  ],
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.4.0",
-    "@typescript-eslint/parser": "^7.4.0",
+    "@eslint/js": "^9.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.11.0",
+    "@typescript-eslint/parser": "^7.11.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.29.1",
@@ -45,14 +37,15 @@
   "devDependencies": {
     "@commitlint/cli": "^17.7.2",
     "@commitlint/config-conventional": "^13.1.0",
-    "eslint": "^8.23.0",
+    "@types/eslint": "^8.56.10",
+    "eslint": "^8.57.0",
     "eslint-find-rules": "^4.1.0",
     "husky": "^7.0.2",
     "semantic-release": "^23.0.6",
-    "typescript": "^4.5.2"
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
-    "eslint": ">=8.23.0",
+    "eslint": ">=8.57.0",
     "typescript": "*"
   }
 }
