--- prettier.json
@@ -1,7 +1,0 @@
-{
-  "printWidth": 120,
-  "trailingComma": "all",
-  "singleQuote": true,
-  "endOfLine": "auto",
-  "arrowParens": "always"
-}
