--- rules/eslint.js
@@ -1,4 +1,0 @@
-module.exports = {
-  'curly': ['error', 'all'],
-  'eqeqeq': ['error', 'smart'],
-};
