--- rules/typescript-eslint.js
@@ -1,179 +1,0 @@
-const { tsiif } = require('../helpers/iif');
-
-module.exports = {
-  '@typescript-eslint/adjacent-overload-signatures': 'error',
-  '@typescript-eslint/array-type': ['error', {
-    default: 'array',
-  }],
-  '@typescript-eslint/await-thenable': 'error',
-  '@typescript-eslint/ban-ts-comment': 'off',
-  '@typescript-eslint/ban-tslint-comment': 'error',
-  '@typescript-eslint/ban-types': 'error',
-  '@typescript-eslint/class-literal-property-style': 'off',
-  '@typescript-eslint/consistent-generic-constructors': 'off',
-  '@typescript-eslint/consistent-indexed-object-style': ['error', 'record'],
-  '@typescript-eslint/consistent-return': 'error', 'consistent-return': 'off',
-  '@typescript-eslint/consistent-type-assertions': 'off',
-  '@typescript-eslint/consistent-type-definitions': 'off',
-  '@typescript-eslint/consistent-type-exports': tsiif('<3.8', 'off', ['error']),
-  '@typescript-eslint/consistent-type-imports': tsiif('<3.8', 'off', ['error', {
-    disallowTypeAnnotations: true,
-    prefer: 'type-imports',
-    fixStyle: 'separate-type-imports',
-  }]),
-  '@typescript-eslint/default-param-last': 'error', 'default-param-last': 'off',
-  '@typescript-eslint/dot-notation': 'error', 'dot-notation': 'off',
-  '@typescript-eslint/explicit-function-return-type': 'error',
-  '@typescript-eslint/explicit-member-accessibility': 'off',
-  '@typescript-eslint/explicit-module-boundary-types': 'off',
-  '@typescript-eslint/init-declarations': 'error', 'init-declarations': 'off',
-  '@typescript-eslint/method-signature-style': ['error', 'property'],
-  '@typescript-eslint/naming-convention': [
-    'error',
-    {
-      selector: 'default',
-      format: ['camelCase'],
-      leadingUnderscore: 'allow',
-      trailingUnderscore: 'allow',
-    },
-    {
-      selector: 'import',
-      format: ['camelCase', 'PascalCase'],
-    },
-    {
-      selector: 'variable',
-      format: ['camelCase', 'UPPER_CASE'],
-      leadingUnderscore: 'allow',
-      trailingUnderscore: 'allow',
-    },
-    {
-      selector: 'typeLike',
-      format: ['PascalCase'],
-    },
-    {
-      selector: ['objectLiteralProperty', 'typeProperty'],
-      format: null,
-    },
-  ],
-  '@typescript-eslint/no-array-delete': 'error',
-  '@typescript-eslint/no-base-to-string': 'error',
-  '@typescript-eslint/no-confusing-non-null-assertion': 'error',
-  '@typescript-eslint/no-confusing-void-expression': 'error',
-  '@typescript-eslint/no-duplicate-enum-values': 'error',
-  '@typescript-eslint/no-duplicate-type-constituents': 'error',
-  '@typescript-eslint/no-dynamic-delete': 'error',
-  '@typescript-eslint/no-empty-function': ['error', {
-    allow: ['arrowFunctions', 'overrideMethods'],
-  }], 'no-empty-function': 'off',
-  '@typescript-eslint/no-empty-interface': 'error',
-  '@typescript-eslint/no-explicit-any': 'error',
-  '@typescript-eslint/no-extra-non-null-assertion': 'error',
-  '@typescript-eslint/no-extraneous-class': ['error', {
-    allowWithDecorator: true,
-  }],
-  '@typescript-eslint/no-floating-promises': ['error', {
-    ignoreVoid: true,
-  }],
-  '@typescript-eslint/no-for-in-array': 'off',
-  '@typescript-eslint/no-implied-eval': 'error', 'no-implied-eval': 'off',
-  '@typescript-eslint/no-import-type-side-effects': 'error',
-  '@typescript-eslint/no-inferrable-types': 'off',
-  '@typescript-eslint/no-invalid-void-type': 'error',
-  '@typescript-eslint/no-loop-func': 'error', 'no-loop-func': 'off',
-  '@typescript-eslint/no-loss-of-precision': 'error', 'no-loss-of-precision': 'off',
-  '@typescript-eslint/no-magic-numbers': ['error', {
-    ignore: [-1, 0, 1, 2, 10, 100, 1000],
-    ignoreArrayIndexes: true,
-    ignoreDefaultValues: true,
-    ignoreEnums: true,
-    ignoreNumericLiteralTypes: true,
-    ignoreReadonlyClassProperties: true,
-  }], 'no-magic-numbers': 'off',
-  '@typescript-eslint/no-meaningless-void-operator': 'off',
-  '@typescript-eslint/no-misused-new': 'error',
-  '@typescript-eslint/no-misused-promises': 'error',
-  '@typescript-eslint/no-mixed-enums': 'error',
-  '@typescript-eslint/no-namespace': 'error',
-  '@typescript-eslint/no-non-null-asserted-nullish-coalescing': 'error',
-  '@typescript-eslint/no-non-null-asserted-optional-chain': 'error',
-  '@typescript-eslint/no-non-null-assertion': 'off',
-  '@typescript-eslint/no-redundant-type-constituents': 'error',
-  '@typescript-eslint/no-require-imports': 'off',
-  '@typescript-eslint/no-shadow': 'error', 'no-shadow': 'off',
-  '@typescript-eslint/no-this-alias': 'error',
-  '@typescript-eslint/no-unnecessary-boolean-literal-compare': 'error',
-  '@typescript-eslint/no-unnecessary-condition': 'error',
-  '@typescript-eslint/no-unnecessary-qualifier': 'error',
-  '@typescript-eslint/no-unnecessary-type-arguments': 'off',
-  '@typescript-eslint/no-unnecessary-type-assertion': 'error',
-  '@typescript-eslint/no-unnecessary-type-constraint': 'error',
-  '@typescript-eslint/no-unsafe-argument': 'error',
-  '@typescript-eslint/no-unsafe-assignment': 'error',
-  '@typescript-eslint/no-unsafe-call': 'error',
-  '@typescript-eslint/no-unsafe-declaration-merging': 'error',
-  '@typescript-eslint/no-unsafe-enum-comparison': 'error',
-  '@typescript-eslint/no-unsafe-member-access': 'error',
-  '@typescript-eslint/no-unsafe-return': 'error',
-  '@typescript-eslint/no-unsafe-unary-minus': 'error',
-  '@typescript-eslint/no-unused-expressions': 'error', 'no-unused-expressions': 'off',
-  '@typescript-eslint/no-unused-vars': ['warn', {
-    argsIgnorePattern: '^_',
-  }], 'no-unused-vars': 'off',
-  '@typescript-eslint/no-use-before-define': 'off', 'no-use-before-define': 'off',
-  '@typescript-eslint/no-useless-constructor': 'error', 'no-useless-constructor': 'off',
-  '@typescript-eslint/no-useless-empty-export': 'error',
-  '@typescript-eslint/no-useless-template-literals': 'error',
-  '@typescript-eslint/no-var-requires': 'error',
-  '@typescript-eslint/non-nullable-type-assertion-style': 'off',
-  '@typescript-eslint/only-throw-error': 'error', 'no-throw-literal': 'off',
-  '@typescript-eslint/parameter-properties': 'off',
-  '@typescript-eslint/prefer-as-const': 'error',
-  '@typescript-eslint/prefer-destructuring': 'off', 'prefer-destructuring': 'off',
-  '@typescript-eslint/prefer-enum-initializers': 'off',
-  '@typescript-eslint/prefer-find': 'off',
-  '@typescript-eslint/prefer-for-of': 'error',
-  '@typescript-eslint/prefer-function-type': 'error',
-  '@typescript-eslint/prefer-includes': 'off',
-  '@typescript-eslint/prefer-literal-enum-member': ['error', {
-    allowBitwiseExpressions: true,
-  }],
-  '@typescript-eslint/prefer-namespace-keyword': 'error',
-  '@typescript-eslint/prefer-nullish-coalescing': 'error',
-  '@typescript-eslint/prefer-optional-chain': 'error',
-  '@typescript-eslint/prefer-promise-reject-errors': 'error', 'prefer-promise-reject-errors': 'off',
-  '@typescript-eslint/prefer-readonly': 'warn',
-  '@typescript-eslint/prefer-readonly-parameter-types': 'off',
-  '@typescript-eslint/prefer-reduce-type-parameter': 'off',
-  '@typescript-eslint/prefer-regexp-exec': 'off',
-  '@typescript-eslint/prefer-return-this-type': 'error',
-  '@typescript-eslint/prefer-string-starts-ends-with': 'error',
-  '@typescript-eslint/prefer-ts-expect-error': 'error',
-  '@typescript-eslint/promise-function-async': 'error',
-  '@typescript-eslint/require-array-sort-compare': 'off',
-  '@typescript-eslint/require-await': 'off', 'require-await': 'off',
-  '@typescript-eslint/restrict-plus-operands': 'error',
-  '@typescript-eslint/restrict-template-expressions': ['error', {
-    allowAny: false,
-    allowArray: false,
-    allowBoolean: true,
-    allowNever: false,
-    allowNullish: true,
-    allowNumber: true,
-    allowRegExp: true,
-  }],
-  '@typescript-eslint/return-await': ['error', 'in-try-catch'], 'no-return-await': 'off',
-  '@typescript-eslint/sort-type-constituents': 'error',
-  '@typescript-eslint/strict-boolean-expressions': ['warn', {
-    allowString: true,
-    allowNumber: true,
-    allowNullableObject: true,
-    allowNullableBoolean: false,
-    allowNullableString: false,
-    allowNullableNumber: false,
-    allowNullableEnum: false,
-    allowAny: false,
-  }],
-  '@typescript-eslint/switch-exhaustiveness-check': 'error',
-  '@typescript-eslint/unbound-method': 'off',
-  '@typescript-eslint/use-unknown-in-catch-callback-variable': 'error',
-}
