--- rules/unicorn.js
@@ -1,8 +1,0 @@
-module.exports = {
-  'unicorn/escape-case': 'error',
-  'unicorn/expiring-todo-comments': ['error', {
-    allowWarningComments: false,
-  }],
-  'unicorn/no-abusive-eslint-disable': 'error',
-  'unicorn/no-console-spaces': 'error',
-}
