+++ src/rules/eslint.ts
@@ -1,0 +1,10 @@
+import { Linter } from 'eslint';
+
+const config: Linter.FlatConfig = {
+  rules: {
+    'curly': ['error', 'all'],
+    'eqeqeq': ['error', 'smart'],
+  },
+};
+
+export default config;
