+++ tsconfig.json
@@ -1,10 +1,16 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "module": "commonjs",
     "lib": ["ES2015"],
-    "strictNullChecks": true
+    "module": "CommonJS",
+    "moduleResolution": "node",
+    "declaration": true,
+    "strictNullChecks": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "noEmit": true,
+    "skipLibCheck": true
   },
-  "include": ["test/*.ts"],
+  "include": ["src", "test"],
   "exclude": ["node_modules"]
 }
